# openapi.json
{
  "info": {
    "title": "cacheppuccino",
    "version": "dev"
  },
  "openapi": "3.0.3",
  "paths": {
    "/cluster": {
      "get": {
        "operationId": "getCluster",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "nullable": true,
                      "properties": {
                        "alarms": {
                          "items": {
                            "properties": {
                              "detail": {
                                "type": "string"
                              },
                              "firing": {
                                "type": "boolean"
                              },
                              "name": {
                                "type": "string"
                              },
                              "since": {
                                "type": "string"
                              }
                            },
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "answered_by": {
                          "type": "string"
                        },
                        "apps": {
                          "items": {
                            "properties": {
                              "app": {
                                "type": "string"
                              },
                              "hashes": {
                                "items": {
                                  "type": "string"
                                },
                                "type": "array"
                              },
                              "in_sync": {
                                "type": "boolean"
                              },
                              "newest_age_seconds": {
                                "format": "int64",
                                "type": "integer"
                              },
                              "rows": {
                                "type": "integer"
                              },
                              "servable_pods": {
                                "type": "integer"
                              }
                            },
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "degraded": {
                          "type": "boolean"
                        },
                        "in_sync": {
                          "type": "boolean"
                        },
                        "pod_count": {
                          "type": "integer"
                        },
                        "pods": {
                          "items": {
                            "properties": {
                              "addr": {
                                "type": "string"
                              },
                              "apps": {
                                "items": {
                                  "properties": {
                                    "app": {
                                      "type": "string"
                                    },
                                    "consecutive_failures": {
                                      "type": "integer"
                                    },
                                    "hash": {
                                      "type": "string"
                                    },
                                    "imported_at": {
                                      "type": "string"
                                    },
                                    "last_pull_error": {
                                      "type": "string"
                                    },
                                    "rows": {
                                      "type": "integer"
                                    },
                                    "servable": {
                                      "type": "boolean"
                                    }
                                  },
                                  "type": "object"
                                },
                                "type": "array"
                              },
                              "pod": {
                                "type": "string"
                              },
                              "primary": {
                                "type": "boolean"
                              },
                              "ready": {
                                "type": "boolean"
                              },
                              "version": {
                                "type": "string"
                              }
                            },
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "primary": {
                          "type": "string"
                        },
                        "ready_count": {
                          "type": "integer"
                        },
                        "unreachable": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        }
                      },
                      "type": "object"
                    },
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "description": ""
          }
        },
        "summary": "Fleet view: every pod, the current primary, and whether replicas agree"
      }
    },
...
    "/monitor": {
      "get": {
        "operationId": "getMonitor",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "nullable": true,
                      "properties": {
                        "alarms": {
                          "items": {
                            "properties": {
                              "detail": {
                                "type": "string"
                              },
                              "firing": {
                                "type": "boolean"
                              },
                              "name": {
                                "type": "string"
                              },
                              "since": {
                                "type": "string"
                              }
                            },
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "all_alarms": {
                          "items": {
                            "properties": {
                              "detail": {
                                "type": "string"
                              },
                              "firing": {
                                "type": "boolean"
                              },
                              "name": {
                                "type": "string"
                              },
                              "since": {
                                "type": "string"
                              }
                            },
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "answered_by": {
                          "type": "string"
                        },
                        "in_sync": {
                          "type": "boolean"
                        },
                        "pod_count": {
                          "type": "integer"
                        },
                        "ready_count": {
                          "type": "integer"
                        },
                        "status": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    },
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "No alarms firing"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "nullable": true,
                      "properties": {
                        "alarms": {
                          "items": {
                            "properties": {
                              "detail": {
                                "type": "string"
                              },
                              "firing": {
                                "type": "boolean"
                              },
                              "name": {
                                "type": "string"
                              },
                              "since": {
                                "type": "string"
                              }
                            },
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "all_alarms": {
                          "items": {
                            "properties": {
                              "detail": {
                                "type": "string"
                              },
                              "firing": {
                                "type": "boolean"
                              },
                              "name": {
                                "type": "string"
                              },
                              "since": {
                                "type": "string"
                              }
                            },
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "answered_by": {
                          "type": "string"
                        },
                        "in_sync": {
                          "type": "boolean"
                        },
                        "pod_count": {
                          "type": "integer"
                        },
                        "ready_count": {
                          "type": "integer"
                        },
                        "status": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    },
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "At least one fleet alarm is firing"
          },
          "default": {
            "description": ""
          }
        },
        "summary": "Fleet health for an external uptime check; 503 when an alarm is firing"
      }
    },
    "/openapi.json": {
      "get": {
        "operationId": "getOpenAPI",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "description": ""
          }
        },
        "summary": "OpenAPI schema"
      }
    },
...
    "/status": {
      "get": {
        "operationId": "getStatus",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "nullable": true,
                      "properties": {
                        "apps": {
                          "items": {
                            "properties": {
                              "age_seconds": {
                                "format": "int64",
                                "type": "integer"
                              },
                              "app": {
                                "type": "string"
                              },
                              "consecutive_failures": {
                                "type": "integer"
                              },
                              "hash": {
                                "type": "string"
                              },
                              "imported_at": {
                                "type": "string"
                              },
                              "last_pull": {
                                "type": "string"
                              },
                              "last_pull_error": {
                                "type": "string"
                              },
                              "rows": {
                                "type": "integer"
                              },
                              "servable": {
                                "type": "boolean"
                              },
                              "source": {
                                "type": "string"
                              }
                            },
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "pod": {
                          "type": "string"
                        },
                        "primary": {
                          "type": "boolean"
                        },
                        "primary_holder": {
                          "type": "string"
                        },
                        "ready": {
                          "type": "boolean"
                        },
                        "uptime_seconds": {
                          "format": "int64",
                          "type": "integer"
                        },
                        "version": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    },
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "description": ""
          }
        },
        "summary": "This pod's status and per-application sync state"
      }
    },
    "/strings": {
      "get": {
        "operationId": "getStrings",
        "parameters": [
          {
            "description": "Application id. Defaults to default.",
            "example": "default",
            "in": "query",
            "name": "app",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Language code (e.g. en, fr, es).",
            "example": "en",
            "in": "query",
            "name": "lang",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Repeatable page param. Example: /strings?page=home\u0026page=about\u0026lang=en",
            "example": "home",
            "explode": true,
            "in": "query",
            "name": "page",
            "schema": {
              "type": "string"
            },
            "style": "form"
          },
          {
            "description": "Comma-separated pages. Example: /strings?pages=home,about\u0026lang=en",
            "example": "home,about",
            "in": "query",
            "name": "pages",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "nullable": true,
                      "properties": {
                        "app": {
                          "type": "string"
                        },
                        "lang": {
                          "type": "string"
                        },
                        "pages": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "strings": {
                          "additionalProperties": {
                            "additionalProperties": {
                              "type": "string"
                            },
                            "type": "object"
                          },
                          "type": "object"
                        }
                      },
                      "type": "object"
                    },
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "304": {
            "description": "Not modified (If-None-Match matched the current ETag)"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "nullable": true,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "details": {
                          "additionalProperties": {
                            "type": "string"
                          },
                          "type": "object"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    },
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "nullable": true,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "details": {
                          "additionalProperties": {
                            "type": "string"
                          },
                          "type": "object"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    },
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Unknown application"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "nullable": true,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "details": {
                          "additionalProperties": {
                            "type": "string"
                          },
                          "type": "object"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    },
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "No translations loaded for this application"
          },
          "default": {
            "description": ""
          }
        },
        "summary": "Get strings for one or more pages in a given language"
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}